// frontend/src/api/generated/v1/models/plannerContractServiceTierEnum.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Wedding Management API (Ninja)
 * OpenAPI spec version: 1.0.0
 */

export type PlannerContractServiceTierEnum = typeof PlannerContractServiceTierEnum[keyof typeof PlannerContractServiceTierEnum];


export const PlannerContractServiceTierEnum = {
  COMPLETA: 'COMPLETA',
  PARCIAL: 'PARCIAL',
  FINAL: 'FINAL',
} as const;