// frontend/src/api/generated/v1/models/plannerContractStatusEnum.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Wedding Management API (Ninja)
 * OpenAPI spec version: 1.0.0
 */

export type PlannerContractStatusEnum = typeof PlannerContractStatusEnum[keyof typeof PlannerContractStatusEnum];


export const PlannerContractStatusEnum = {
  DRAFT: 'DRAFT',
  SIGNED: 'SIGNED',
  CANCELED: 'CANCELED',
} as const;